Add files via upload
Project: Math Game In C++ With Source Code
Please scroll down and click on the download button to download Math Game In C++ With Source Code for free

The Math Game is a single-player game. This project is an interesting project. Here, the user can test their calculation skills through this app. If you are able to give the right solution to the problem, then you will get a score of 10 on each question. If not, then the system will end the game. This whole system is in the C++ language.

Making of the game
This Math game is simply designed in ‘C++’ language. Taking about the features of this system, the user can test their calculation skills through this game. You just have to type the answer within 5 seconds and press enter to check the answer. If you are right, then you will get a score of 10. If not, then the system will end the game. The game score is shown at last with a high score. Also, this project is easy to operate and understood by the users.

How to run the project?
In order to run the project, you must have installed Dev C++ or Code Blocks on your PC. (i recommend these IDE). The Math Game in C++ with source code is free to download,